requirements.txt
asgiref==3.2.10
bootstrap4==0.1.0
bson==0.5.8
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
coverage==5.0.4
cryptography==3.0
dataclasses==0.6
Django==2.2.10
djangorestframework==3.11.1
djongo==1.3.1
fire==0.3.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
six==1.14.0
sqlparse==0.2.4
termcolor==1.1.0
urllib3==1.25.10
uvloop==0.14.0
websockets==8.1
whitenoise==5.2.0
